+++ src/main/java/com/bcafinance/engineapproval/model/ParamField.java
@@ -13,7 +13,7 @@ import javax.persistence.Table;
 @Data @Entity @NoArgsConstructor @AllArgsConstructor
 @Table(name = "SP_ME_Approval_Param_Field")
 public class ParamField {
-    private @Id @GeneratedValue(strategy=GenerationType.IDENTITY) Long id;
+    private @Id @GeneratedValue(strategy=GenerationType.IDENTITY) int id;
     private String fieldText;
     private String fieldVal;
     private String fieldSourceTable;
