+++ src/main/java/com/bcafinance/engineapproval/model/TempDetailTemplate.java
@@ -13,7 +13,7 @@ import javax.persistence.Table;
 @Data @Entity @NoArgsConstructor @AllArgsConstructor
 @Table(name = "SP_ME_Approval_DetailTemplate")
 public class TempDetailTemplate {
-    private @Id @GeneratedValue(strategy=GenerationType.IDENTITY) Long id;
+    private @Id @GeneratedValue(strategy=GenerationType.IDENTITY) int id;
     private Long master_id;
     private String idx;
     private String operator;
